Move ThresholdSummaryChip to threhsolds directory
